# sql: use binary search to fetch in JSON objects