@jordanlewis 2018-01-22
LGTM! Could you add comments to the top of each of the new functions you've added? I think I get the idea from their nicely descriptive names, but I think a little precision in what exactly they're returning would go a long way.